Not inovoke nyoibo when invoke rails by rake or test mode.

+++ lib/nyoibo/railtie.rb
@@ -3,12 +3,14 @@ require "nyoibo"
 class Nyoibo::Railtie < Rails::Railtie
   config.after_initialize do
     ENV["NYOIBO_ENV"] = Rails.env
-    Nyoibo.run
-    ::ApplicationController.send(:prepend_before_filter, lambda{
-                                   if Rails.env == "development"
-                                     Process.kill(:INT, Nyoibo.pid)
-                                     Nyoibo.run
-                                   end
-                                   true})
+    if Rails.env != "test" && !defined?(::Rake)
+      Nyoibo.run
+      ::ApplicationController.send(:prepend_before_filter, lambda{
+                                     if Rails.env == "development"
+                                       Process.kill(:INT, Nyoibo.pid)
+                                       Nyoibo.run
+                                     end
+                                     true})
+    end
   end
 end
